+++ src/go/replace-conflicts.go
@@ -16,7 +16,7 @@ func ReplaceConflicts(stdOut io.Writer, confirmed bool) {
 	checkConfirmed(stdOut, confirmed)
 	ex.ExecuteOrDie(ex.ExecuteOptions{}, "git", "reset", "--hard", "HEAD")
 	slog.Info(fmt.Sprint("Replacing changes (merge conflicts) for failed rebase of commit ", commitWithConflicts, ", with changes from associated branch, ", branchInfo.BranchName))
-	diff := ex.ExecuteOrDie(ex.ExecuteOptions{}, "git", "diff", "--binary", "origin/"+GetMainBranch(), branchInfo.BranchName)
+	diff := ex.ExecuteOrDie(ex.ExecuteOptions{}, "git", "diff", "--binary", "origin/"+GetMainBranchOrDie(), branchInfo.BranchName)
 	ex.ExecuteOrDie(ex.ExecuteOptions{Stdin: &diff}, "git", "apply")
 	slog.Info("Adding changes and continuing rebase")
 	ex.ExecuteOrDie(ex.ExecuteOptions{}, "git", "add", ".")