+++ src/go/sd/command_rebase_main.go
@@ -11,7 +11,7 @@ func CreateRebaseMainCommand() Command {
 	return Command{
 		FlagSet: flagSet,
 		Summary: "Bring your main branch up to date with remote",
-		Description: "Rebase with origin/" + sd.GetMainBranch() + ", dropping any commits who's associated\n" +
+		Description: "Rebase with origin/" + sd.GetMainBranchForHelp() + ", dropping any commits who's associated\n" +
 			"branches have been merged.\n" +
 			"\n" +
 			"This avoids having to manually call \"git reset --hard head\" whenever\n" +