+++ src/go/sd/command_replace_commit.go
@@ -10,14 +10,14 @@ func CreateReplaceCommitCommand() Command {
 	indicatorTypeString := AddIndicatorFlag(flagSet)
 	return Command{
 		FlagSet: flagSet,
-		Summary: "Replaces a commit on " + sd.GetMainBranch() + " branch with its associated branch",
-		Description: "Replaces a commit on " + sd.GetMainBranch() + " branch with the squashed contents of its\n" +
+		Summary: "Replaces a commit on " + sd.GetMainBranchForHelp() + " branch with its associated branch",
+		Description: "Replaces a commit on " + sd.GetMainBranchForHelp() + " branch with the squashed contents of its\n" +
 			"associated branch.\n" +
 			"\n" +
 			"This is useful when you make changes within a branch, for example to\n" +
 			"fix a problem found on CI, and want to bring the changes over to your\n" +
-			"local " + sd.GetMainBranch() + " branch.",
-		Usage: "sd replace-commit [flags] <commitIndicator>",
+			"local " + sd.GetMainBranchForHelp() + " branch.",
+		Usage: "sd " + flagSet.Name() + " [flags] <commitIndicator>",
 		OnSelected: func(command Command) {
 			if flagSet.NArg() == 0 {
 				commandError(flagSet, "missing commitIndicator", command.Usage)