+++ src/go/sd/command_replace_conflicts.go
@@ -14,8 +14,8 @@ func CreateReplaceConflictsCommand(stdOut io.Writer) Command {
 		Summary: "For failed rebase: replace changes with its associated branch",
 		Description: "During a rebase that failed because of merge conflicts, replace the\n" +
 			"current uncommitted changes (merge conflicts), with the contents\n" +
-			"(diff between origin/" + sd.GetMainBranch() + " and HEAD) of its associated branch.",
-		Usage: "sd replace-conflicts",
+			"(diff between origin/" + sd.GetMainBranchForHelp() + " and HEAD) of its associated branch.",
+		Usage: "sd " + flagSet.Name(),
 		OnSelected: func(command Command) {
 			if flagSet.NArg() > 0 {
 				commandError(flagSet, "too many arguments", command.Usage)